App: move creationg of the app object inside the on_startup callback
We can't create application windows before we connect to the
startup singal, which means we currently create an App object
with optionals and refcell wrappers and initialize it after wards.

Moving everything inside the on_startup signal will allow use to
remove the optionals from the AppInner struct and initialize
everything from the start.